Address code review: fix proxy naming, type precision, and magic number
Agent-Logs-Url: https://github.com/ikkentim/SampSharp/sessions/ebe1fede-b4b3-47ab-8f30-762f92b8f853

+++ src/SampSharp.OpenMp.Entities/SAMP/Systems/NpcSystem.cs
@@ -145,7 +145,7 @@ internal class NpcSystem : DisposableSystem, INPCEventHandler
         _eventDispatcher.Invoke("OnNPCFinishNodePoint",
             _entityProvider.GetEntity(npc),
             nodeId,
-            (int)pointId);
+            pointId);
     }
 
     public void OnNPCFinishNode(INPC npc, int nodeId)
